Moved log.py to folder - types
