+++ trolley/__init__.py
@@ -7,4 +7,4 @@ from trolley.payment import Payment
 from trolley.recipient_account import RecipientAccount
 from trolley.gateway import Gateway
 from trolley.batch_summary import BatchSummary
-from trolley.log import Log
+from trolley.types.log import Log
